backend/requirements.txt
django
autoflake
bcrypt
mypy
isort
black
rich
attrs
djangorestframework_simplejwt
djangorestframework
markdown
django-filter
